src/Agent_theory/requirements.txt
pandas==2.2.2
numpy==1.26.4
faiss-cpu==1.8.0
sentence-transformers==3.0.1
torch
FlagEmbedding
google-genai